+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Prop `thresholdForFacetSearch` that displays a Search bar to filter facets.
+
 ## [3.69.3] - 2020-09-03
 ### Fixed
 - Prevent `SearchResultFlexible` from overflowing on mobile devices by setting its width.